+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Handle a full list of errors when check out fails. Refs UICHKOUT-679.
 * Item blocks: Allow for override when logged in user has credentials. Refs UICHKOUT-677.
 * Add support for optional `readyPrefix` property at the module level in stripes.config.js. If set, this prefix will be displayed in the title when the app is ready to receive a scanned item barcode. Implements UICHKOUT-686.
+* Remove old override endpoint and permission. Refs UICHKOUT-688.
 
 
 ## [5.0.0](https://github.com/folio-org/ui-checkout/tree/v5.0.0) (2020-10-12)